requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
biopython==1.86
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.6.77
pandas==2.3.3
pillow==12.1.0
propcache==0.4.1
py3Dmol==2.5.3
pyarrow==23.0.0
pyparsing==3.3.2
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.0
tqdm==4.67.1
transformers==4.54.1
triton==3.3.0
tzdata==2025.3
urllib3==2.6.3
xxhash==3.6.0
yarl==1.22.0
